requirements.txt
pygame 
numpy
PyOpenGL
glfw 
numpy
opencv-python
mediapipe